requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via hummingbot-api-client
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via httpx
attrs==25.3.0
    # via aiohttp
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via -r requirements.in
commlib-py==0.11.5
    # via hbotrc
docker==7.1.0
    # via -r requirements.in
filelock==3.19.1
    # via torch
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via torch
h11==0.16.0
    # via httpcore
hbotrc @ git+https://github.com/hummingbot/hbot-remote-client-py.git
    # via -r requirements.in
hiredis==3.2.1
    # via redis
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
hummingbot-api-client==1.1.4
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
llvmlite==0.45.1
    # via numba
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
networkx==3.5
    # via torch
numba==0.62.1
    # via stumpy
numpy==2.3.2
    # via
    #   -r requirements.in
    #   numba
    #   pyro-ppl
    #   scipy
    #   stumpy
    #   torchvision
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.3
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.8.90
    # via torch
opt-einsum==3.4.0
    # via pyro-ppl
paho-mqtt==1.6.1
    # via commlib-py
pika==1.3.2
    # via commlib-py
pillow==11.3.0
    # via torchvision
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via commlib-py
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via redis
pyro-api==0.1.2
    # via pyro-ppl
pyro-ppl==1.9.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.3
    # via -r requirements.in
redis[hiredis]==5.3.1
    # via commlib-py
requests==2.32.4
    # via
    #   commlib-py
    #   docker
rich==13.9.4
    # via
    #   -r requirements.in
    #   commlib-py
scipy==1.16.1
    # via
    #   stumpy
    #   torchdiffeq
sniffio==1.3.1
    # via anyio
stumpy==1.13.0
    # via -r requirements.in
sympy==1.14.0
    # via torch
torch==2.8.0
    # via
    #   -r requirements.in
    #   pyro-ppl
    #   torchaudio
    #   torchdiffeq
    #   torchvision
torchaudio==2.8.0
    # via -r requirements.in
torchdiffeq==0.2.5
    # via -r requirements.in
torchvision==0.23.0
    # via -r requirements.in
tqdm==4.67.1
    # via pyro-ppl
triton==3.4.0
    # via torch
typing-extensions==4.14.1
    # via
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
ujson==5.11.0
    # via commlib-py
urllib3==2.5.0
    # via
    #   docker
    #   requests
wheel==0.45.1
    # via commlib-py
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
